Move feedAction up to feedList component
