+++ src/components/dashboard/FeedActions.js
@@ -2,7 +2,6 @@
 import React from 'react'
 import { TouchableOpacity, View } from 'react-native-web'
 import { withStyles } from '../../lib/styles'
-import goodWallet from '../../lib/wallet/GoodWallet'
 import { Icon, Text } from '../common'
 import type { FeedEventProps } from './FeedItems/EventProps'
 
@@ -12,50 +11,28 @@ import type { FeedEventProps } from './FeedItems/EventProps'
  * @param {FeedEventProps} feedItem - Contains the feed item
  * @returns React element with actions
  */
-const FeedActions = ({ item, styles, theme }: FeedEventProps) => {
-  const canDelete = item && item.id && item.id.indexOf('0x') === -1
-  const canCancel = item && item.displayType === 'sendpending'
+const FeedActions = ({ canCancel, canDelete, children, item, onPress, styles, theme }: FeedEventProps) => {
   const hasAction = canDelete || canCancel
   const backgroundColor = hasAction ? theme.colors.red : 'transparent'
 
-  const handleActionPress = async () => {
-    if (canCancel) {
-      try {
-        await goodWallet.cancelOTLByTransactionHash(item.id)
-      } catch (e) {
-        console.info('error', e)
-      }
-    }
-
-    // TODO: canDelete action
+  const handleActionPress = () => {
+    onPress(item.id, { canCancel, canDelete })
   }
 
   return (
     <View style={[styles.actionsContainer, { backgroundColor }]}>
       {hasAction && (
-        <TouchableOpacity onPress={handleActionPress}>
+        <TouchableOpacity onPress={hasAction && handleActionPress}>
           <View style={styles.actionsContainerInner}>
             <Icon name="close" color={theme.colors.surface} />
             <Text style={[styles.action]} fontSize={14} fontWeight="500" color="surface">
-              {actionLabel({ canDelete, canCancel })}
+              {children}
             </Text>
           </View>
         </TouchableOpacity>
       )}
     </View>
   )
-}
-
-const actionLabel = ({ canDelete, canCancel }) => {
-  if (canCancel) {
-    return 'Cancel Payment Link'
-  }
-
-  if (canDelete) {
-    return 'Delete'
-  }
-
-  return ''
 }
 
 const getStylesFromProps = ({ theme }) => ({
